fix PAGE_MASK

+++ src/config.cc
@@ -96,7 +96,7 @@ int _get_bits_of(int v) {
   return n;
 }
 unsigned _page_size = sysconf(_SC_PAGESIZE);
-unsigned long _page_mask = ~(_page_size - 1);
+unsigned long _page_mask = ~(unsigned long)(_page_size - 1);
 unsigned _page_shift = _get_bits_of(_page_size);
 
 int _num_threads = 0;
